feat: transform the hardcoded params into args
Functions train() and load_data() used to deal with a lot of parameters
that are initialized inside these functions. In order to add some
flexibility, let us convert those parameters into the function arguments.
Then, in order to run the program for different models, there will be no
need to modify the functions. It will be enough to call the functions with
different arguments.